backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.29.0

sqlmodel==0.0.12
psycopg2-binary==2.9.10
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-chess==1.999
stockfish==3.28.0
python-dotenv==1.0.1
email-validator==1.3.1
